+++ package.json
@@ -44,6 +44,7 @@
     "formik": "^2.2.9",
     "graphql": "^16.3.0",
     "lodash": "^4.17.21",
+    "markdown-to-jsx": "^7.1.7",
     "next": "12.1.0",
     "postcss-calc": "^8.2.4",
     "postcss-flexbugs-fixes": "^5.0.2",
