+++ src/components/Icon/IconShareFat/index.tsx
@@ -1,0 +1,4 @@
+import { withIcon } from "../withIcon";
+import { ReactComponent as Icon } from "./share.svg";
+
+export const IconShareFat = withIcon(Icon);
