+++ src/views/Logbook/index.tsx
@@ -1,20 +1,22 @@
-import Link from "next/link";
+import React, { useState, useEffect } from "react";
 import { useRouter } from "next/router";
-import { useEffect } from "react";
 import { useQuery } from "@apollo/client";
+import { useAccount } from "wagmi";
 
-import { Head, Button, Spinner } from "~/components";
+import { Head, Spinner } from "~/components";
 import { LOGBOOK_DETAIL } from "~/components/GQL";
 
-import { Nav } from "./Nav";
-import { EmptyBook } from "./EmptyBook";
-import LogbookDetail from "./LogbookDetail";
+import { Book } from "./Book";
+import Editor from "./Editor";
 
 import styles from "./styles.module.css";
 
 const Logbook: React.FC = () => {
+  const [isEditing, enableEditing] = useState(false);
+  const [{ data: accountData }] = useAccount();
   const router = useRouter();
   const id = router.query.id as string;
+  const tokenID = `0x${Number(id).toString(16)}`;
 
   const {
     loading,
@@ -23,7 +25,7 @@ const Logbook: React.FC = () => {
     // fetchMore,
   } = useQuery(LOGBOOK_DETAIL, {
     variables: {
-      id,
+      id: tokenID,
     },
   });
 
@@ -31,37 +33,43 @@ const Logbook: React.FC = () => {
     console.log("logbookDetail:", logbookDetail);
   }, [logbookDetail]);
 
+  if (loading) {
+    return <Spinner />;
+  }
+
+  if (!logbookDetail?.logbook) {
+    return <></>;
+  }
+
+  const {
+    logbook: { title, transferCount, description, publications, owner },
+  } = logbookDetail;
+
   return (
     <>
       <Head title="Logbook" />
-
-      {/* <h1>Logbook Detail</h1>
-      <Button />
-
-      <ul>
-        <li>
-          <Link href="/">Homepage</Link>
-        </li>
-        <li>
-          <Link href="library">Library</Link>
-        </li>
-        <li>
-          <Link href="bookcase">Bookcase</Link>
-        </li>
-      </ul> */}
-
       <section className={styles.maxWidth}>
-        {router.query.testEditor && logbookDetail ? (
-          <LogbookDetail
-            id={logbookDetail.id}
-            transferCount={logbookDetail.transferCount}
-            content={logbookDetail.publications?.[0]?.log?.content}
-          />
+        {accountData &&
+        accountData?.address.toLowerCase() === owner?.id.toLowerCase() &&
+        isEditing ? (
+          <Editor
+            id={tokenID}
+            transferCount={transferCount}
+            content={publications?.[0]?.log?.content}
+            onLeave={() => enableEditing(false)}
+          ></Editor>
         ) : (
-          <>
-            <EmptyBook tokenID={id} />
-            {loading && <Spinner />}
-          </>
+          <Book
+            tokenID={id}
+            title={title}
+            transferCount={transferCount}
+            description={description}
+            publications={publications}
+            isOwn={
+              accountData?.address.toLowerCase() === owner?.id.toLowerCase()
+            }
+            onEdit={() => enableEditing(true)}
+          />
         )}
       </section>
     </>
